# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "BrowseQuery": {
        "description": "Normalized public criteria echoed with browse results.",
        "properties": {
          "direction": {
            "const": "desc",
            "default": "desc",
            "title": "Direction",
            "type": "string"
          },
          "sort": {
            "const": "popularity",
            "default": "popularity",
            "title": "Sort",
            "type": "string"
          }
        },
        "title": "BrowseQuery",
        "type": "object"
      },
      "CatalogOption": {
        "description": "Stable identifier and English source label for one filter option.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label"
        ],
        "title": "CatalogOption",
        "type": "object"
      },
      "DurationKind": {
        "description": "Supported campaign-duration measures.",
        "enum": [
          "fast",
          "normal",
          "completionist"
        ],
        "title": "DurationKind",
        "type": "string"
      },
      "ErrorCode": {
        "description": "Stable failure codes understood by public adapters.",
        "enum": [
          "GAME_NOT_FOUND",
          "INTERNAL_ERROR",
          "INVALID_QUERY",
          "METHOD_NOT_ALLOWED",
          "NOT_FOUND",
          "RATE_LIMITED",
          "UPSTREAM_INVALID_RESPONSE",
          "UPSTREAM_TIMEOUT",
          "UPSTREAM_UNAVAILABLE",
          "VALIDATION_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "description": "Public details nested inside an error response.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "retryAfterSeconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryafterseconds"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Public envelope shared by every error response.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FilterLimits": {
        "description": "Public bounds shared by filter controls and query validation.",
        "properties": {
          "maximumDurationHours": {
            "title": "Maximumdurationhours",
            "type": "integer"
          },
          "maximumNameLength": {
            "title": "Maximumnamelength",
            "type": "integer"
          },
          "maximumPage": {
            "title": "Maximumpage",
            "type": "integer"
          },
          "minimumAutocompleteLength": {
            "title": "Minimumautocompletelength",
            "type": "integer"
          },
          "minimumDurationHours": {
            "title": "Minimumdurationhours",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          }
        },
        "required": [
          "pageSize",
          "maximumPage",
          "minimumAutocompleteLength",
          "maximumNameLength",
          "minimumDurationHours",
          "maximumDurationHours"
        ],
        "title": "FilterLimits",
        "type": "object"
      },
      "FilterMetadata": {
        "description": "All allow-listed values needed to construct the search form.",
        "properties": {
          "durationKinds": {
            "items": {
              "$ref": "#/components/schemas/DurationKind"
            },
            "title": "Durationkinds",
            "type": "array"
          },
          "gameModes": {
            "items": {
              "$ref": "#/components/schemas/CatalogOption"
            },
            "title": "Gamemodes",
            "type": "array"
          },
          "genres": {
            "items": {
              "$ref": "#/components/schemas/CatalogOption"
            },
            "title": "Genres",
            "type": "array"
          },
          "limits": {
            "$ref": "#/components/schemas/FilterLimits"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/CatalogOption"
            },
            "title": "Platforms",
            "type": "array"
          },
          "sortOptions": {
            "items": {
              "$ref": "#/components/schemas/SortOption"
            },
            "title": "Sortoptions",
            "type": "array"
          }
        },
        "required": [
          "platforms",
          "genres",
          "gameModes",
          "durationKinds",
          "sortOptions",
          "limits"
        ],
        "title": "FilterMetadata",
        "type": "object"
      },
      "GameCover": {
        "description": "Normalized game cover image.",
        "properties": {
          "height": {
            "exclusiveMinimum": 0.0,
            "title": "Height",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "width": {
            "exclusiveMinimum": 0.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "url",
          "width",
          "height"
        ],
        "title": "GameCover",
        "type": "object"
      },
      "GamePage": {
        "description": "One normalized page of game summaries.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GameSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "query": {
            "$ref": "#/components/schemas/BrowseQuery"
          }
        },
        "required": [
          "items",
          "pagination",
          "query",
          "meta"
        ],
        "title": "GamePage",
        "type": "object"
      },
      "GameRating": {
        "description": "Normalized rating and the number of contributing scores.",
        "properties": {
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "value": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "value",
          "count",
          "source"
        ],
        "title": "GameRating",
        "type": "object"
      },
      "GameSummary": {
        "description": "Provider-neutral game data required by browse results.",
        "properties": {
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameCover"
              },
              {
                "type": "null"
              }
            ]
          },
          "gameModes": {
            "items": {
              "$ref": "#/components/schemas/CatalogOption"
            },
            "title": "Gamemodes",
            "type": "array"
          },
          "genres": {
            "items": {
              "$ref": "#/components/schemas/CatalogOption"
            },
            "title": "Genres",
            "type": "array"
          },
          "id": {
            "exclusiveMinimum": 0.0,
            "title": "Id",
            "type": "integer"
          },
          "normalDurationSeconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normaldurationseconds"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/CatalogOption"
            },
            "title": "Platforms",
            "type": "array"
          },
          "rating": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameRating"
              },
              {
                "type": "null"
              }
            ]
          },
          "releaseYear": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Releaseyear"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "releaseYear",
          "cover",
          "platforms",
          "genres",
          "rating",
          "normalDurationSeconds",
          "gameModes"
        ],
        "title": "GameSummary",
        "type": "object"
      },
      "HealthResponse": {
        "description": "Stable public response for process-health checks.",
        "properties": {
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "HealthResponse",
        "type": "object"
      },
      "Pagination": {
        "description": "Page location and total result information.",
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "const": 24,
            "default": 24,
            "title": "Pagesize",
            "type": "integer"
          },
          "totalItems": {
            "minimum": 0.0,
            "title": "Totalitems",
            "type": "integer"
          },
          "totalPages": {
            "minimum": 0.0,
            "title": "Totalpages",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "totalItems",
          "totalPages"
        ],
        "title": "Pagination",
        "type": "object"
      },
      "ResponseMeta": {
        "description": "Freshness and processing metadata for one catalog response.",
        "properties": {
          "dataMayBeStale": {
            "default": false,
            "title": "Datamaybestale",
            "type": "boolean"
          },
          "excludedUnknownDuration": {
            "default": false,
            "title": "Excludedunknownduration",
            "type": "boolean"
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid"
          },
          "servedFrom": {
            "$ref": "#/components/schemas/ServedFrom",
            "default": "provider"
          }
        },
        "title": "ResponseMeta",
        "type": "object"
      },
      "ServedFrom": {
        "description": "Origin of the data returned to the caller.",
        "enum": [
          "provider"
        ],
        "title": "ServedFrom",
        "type": "string"
      },
      "SortOption": {
        "description": "Supported catalog sort fields.",
        "enum": [
          "popularity",
          "rating",
          "release-date",
          "duration",
          "title"
        ],
        "title": "SortOption",
        "type": "string"
      }
    }
  },
  "info": {
    "description": "Provider-neutral game discovery API.",
    "title": "What To Play Next API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/filters": {
      "get": {
        "description": "Return stable filter options from the configured catalog.",
        "operationId": "getFilters",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterMetadata"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too many requests.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An unexpected error occurred.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Game data provider returned an invalid response.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Game data is temporarily unavailable.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Game data provider timed out.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get catalog filter metadata",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/games": {
      "get": {
        "description": "Return one unfiltered page ordered by current popularity.",
        "operationId": "browseGames",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 100,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamePage"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "One or more parameter values are invalid.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too many requests.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An unexpected error occurred.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Game data provider returned an invalid response.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Game data is temporarily unavailable.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Game data provider timed out.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Browse games",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "description": "Report liveness without synchronously calling IGDB or Redis.",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An unexpected error occurred.",
            "headers": {
              "X-Request-ID": {
                "description": "Identifier used to correlate this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Check process health",
        "tags": [
          "system"
        ]
      }
    }
  }
}